+++ src/main/kotlin/io/openfuture/chain/service/Services.kt
@@ -1,7 +1,10 @@
 package io.openfuture.chain.service
 
+import io.openfuture.chain.crypto.domain.ECKey
+import io.openfuture.chain.crypto.domain.ExtendedKey
 import io.openfuture.chain.domain.HardwareInfo
-import io.openfuture.chain.domain.block.BlockDto
+import io.openfuture.chain.domain.block.BlockRequest
+import io.openfuture.chain.domain.crypto.key.WalletDto
 import io.openfuture.chain.domain.hardware.CpuInfo
 import io.openfuture.chain.domain.hardware.NetworkInfo
 import io.openfuture.chain.domain.hardware.RamInfo
@@ -42,6 +45,20 @@ interface BlockService {
 interface CryptoService {
 
     fun generateSeedPhrase(): String
+
+    fun getMasterKey(seedPhrase: String): ExtendedKey
+
+    fun getDerivationKey(seedPhrase: String, derivationPath: String): ExtendedKey
+
+    fun importKey(key: String): ExtendedKey
+
+    fun importWifKey(wifKey: String): ECKey
+
+    fun serializePublicKey(key: ExtendedKey): String
+
+    fun serializePrivateKey(key: ExtendedKey): String
+
+    fun generateKey(): WalletDto
 
 }